Revert "gitlab_runner: Initial the keyring in the base image for faster boot"
This reverts commit 466230e41e21fbabeb922909ec226fc6f4d8bab3.

This has been fixed in pacman[1], so it is no longer unreasonably slow.
Some quick testing at runner1 indicates that this only saves five
seconds at best, so IMO it is not worth the complexity to continue doing
this.

This revert does not revert the timeout back to 60 seconds, but keeps it
at 30 seconds.

[1] https://gitlab.archlinux.org/pacman/pacman/-/merge_requests/16