gitlab_runner: Switch to new libvirt-executor image[1] from arch-boxes
It makes more sense to build the image in arch-boxes than building it on
each runner, especially considering that arch-boxes already have all the
necessary infrastructure, so we can avoid maintaining similar code in
two repositories and avoid running losetup, mount, arch-chroot etc. (as
root) on the runners.

The arch-boxes MR[1] has a little more context.

[1] https://gitlab.archlinux.org/archlinux/arch-boxes/-/merge_requests/200