Add server for the nvchecker PoC[1]
The server will be used for developing and running the nvchekcer PoC.
Antiz is working on the nvchecker PoC and will be provided root access.

As it is used for development purposes and we are providing root access
to a person outside the DevOps team, it is not managed by the DevOps
team, but by the nvchecker-poc group (which includes two DevOps
members), and only has the minimum of roles deployed.

It is expected that this will be put into "production" at some point and
thus must be fully managed by the DevOps team. It must be evaluated
before this point whether access to people outside DevOps team should be
provided, and if so, how and under what form of governance.

A playbook is added to ease development of the relevant roles at a later
point.

[1] https://gitlab.archlinux.org/archlinux/nvchecker-poc