tf-stage1: Add GitLab Pages for panic form
This allows users of the 'linux' package to make use of the
DRM-subsystems panic screen that can encode a zlib-encoded panic trace
in a QR code and this way allow a user to transport data via QR to their
computer.

An example URL is the following:

    https://panic.archlinux.org/panic_report/?a=x86_64&v=6.10.0&zl=232110691174483706575043650521221445645717046796552248644098439502487

This site will most likely be extended in the future under it's repo in
the archlinux namespace.

Link: https://gitlab.archlinux.org/archlinux/panic_report
Link: https://gitlab.archlinux.org/archlinux/packaging/packages/linux/-/commit/9faea7f99edb365734ec8d67437da93c3e1a67d1
Signed-off-by: Christian Heusel <christian@heusel.eu>