tf-stage1: Rescale wiki to cpx31
This host was previously a cx31 and was temporarily scaled up to cpx41
to deal with the additional load of a scrape/DDoS. Now we can not scale
it back down to cx31, as this server type is not available anymore in
our location. Additionaly the wiki is crucial to our project, and a few
bucks extra for it shouldn't worry us.

Signed-off-by: Christian Heusel <christian@heusel.eu>