Add opt-in Anubis for the wiki to combat scraping
The wiki has been brought down by scraping and Anubis has proven
effective at blocking this type of attack. The old challenging is
removed as it is mostly ineffective.

If is still undecided if we want Anubis permanently enabled, so for now
it is disabled by default but can quickly be enabled if needed.