archwiki: Set memcached size to be half the machine's total RAM
I don't think the static assignment really makes as it made the machine swap heavily when it was set to 6GB.
Setting it to half the machine's RAM makes sense: It allows the rest to be used for mariadb and IO cache.
Effectively, this will result in the other half being used as mostly IO cache which will in turn be used to
serve NGINX cache.