gitlab_runner: Decrease concurrency to 8
A concurrency of 100 doesn't really make sense for us.
Let's instead go with 8 which seems like a reasonable number.
I picked 8 because it should be neither limiting to us nor
should it be large enough to overwhelm a runner.

If at some point we feel fancy, we could make it dynamic depending
on the machine but for now this should be fine.