get_key.py: Fix setting the default flag
Due to a behavior change in click it does no longer set 'bare' as a
default flag for the `--format` argument when upgrading to a version
post '8.2.0'. Unfortunately we also can not just set `is_flag=True` as
we have a selection of options to work around this.

Without this change the helper script for the accountsupport helpdesk
breaks.

Link: https://github.com/pallets/click/pull/2829
Signed-off-by: Christian Heusel <christian@heusel.eu>