gitlab: Put GitLab behind our own nginx for increased flexibility
We may need to put GitLab behind Anubis in the future, so to be ready
for that scenario we put GitLab behind our own nginx.

GitLab Pages requests are handled directly by the gitlab-pages daemon[1]
on dedicated IP addresses, which is the reason why nginx must listen on
explicit addresses (nginx listening on 0.0.0.0:80 and gitlab-pages
listening on <dedicated-ip>:80 is not possible).

[1] https://gitlab.com/gitlab-org/gitlab-pages