fail2ban: go aggressive and maximum resource sparing with nginx-limit-req
This change can be revisited in a later stage, for the time being we
should prefer maximum resource sparing and minimal "time to action".

The error.log only contains abusers anyway as nginx already did the
counting before adding entries to the log. In best case we can relax
this slightly so offenders will first get graceful http deny codes for a
certain time, but right now the priority is to reduce resources,
time-to-action while maximizing bantime to avoid repeating abusers once
a day. Subsequently also raise the bantime to 1 week.