fail2ban: permanently store entries in firewalld to survive reload
The runtime and the permanent sets are distinct, which means if we only
keep adding them to the runtime ipset but later on need to restart or
reload firewalld, it means all entries are lost. This is quite a problem
as we have a diverged stateful ban between firewalld and fail2ban, which
leads to fail2ban thinking an ip is already banned while firewalld just
doesn't have the entry anymore.

Avoid this loophole by also adding all entries to the permanent set and
subsequently remove the window of opportunity for a state mismatch due
to reloads. Let fail2ban handle the stateful decision if an ip should be
removed or not, while firewalld should make sure it isn't silently
discarded without fail2ban being aware of it.