tf-stage2: Fix fastly returning HTTP status 501
After a lengthy investigation we figured out that the If-Modified-Since
header and respectively returned partial content seems to cause some
issues with the caching logic within fastly.

We therefore rename the headers between origin and shield as suggested
by fastly support while they investigate the issue.

Later on we figured out a totally different fix, which was also
suggested by the fastly support:

> I noticed that your service is setting "pass" to bypass caching for
> certain URLs in cache settings. I would suggest to move those settings
> to request settings when condition to set the "pass" is URL or something
> from client side. "pass" in request settings will work more efficiently
> than cache settings in most cases, and it also can prevent the 501 with
> segmented caching because it will bypass cache lookup. Please see below
> link for more details.

Link: https://www.fastly.com/documentation/guides/full-site-delivery/fastly-vcl/vcl-best-practices/#choose-the-right-pass-behavior-for-your-use-case
Signed-off-by: Christian Heusel <christian@heusel.eu>