archwiki: Write Mediawiki specific anubis rules
This commit enables anubis for the Arch Linux Wiki, but mostly just lets
it forward the requests, except for a a few select endpoints that have
proven to be problematic since they're often crawled and expensive.

We do not want to use the Anubis challenge for every page view, in order
to keep a generally good user experience for our service.

Signed-off-by: Christian Heusel <christian@heusel.eu>