Switch to nginx native ACME[1] where it makes sense
It provides a better UX compared to certbot and is one less moving part.

The switch is done where it is possible and sensible. That is basically
all the places where nginx is the sole user of the certificate.

Certificates using the DNS-01 challenge (e.g. geo mirrors) or which are
used with stream (e.g. rsync over TLS) still require certbot as it is
currently not supported in the native integration[2][3].

[1] https://blog.nginx.org/blog/native-support-for-acme-protocol
[2] https://github.com/nginx/nginx-acme/issues/11
[3] https://github.com/nginx/nginx-acme/issues/13